pyproject.toml
[project]
name = "dio-desafio"
version = "0.1.0"
description = ""
authors = [
    {name = "Lailson Conceicao",email = "lailsondev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "databases[sqlite] (>=0.9.0,<0.10.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
